refactor bin fns to lib
